Simplify purge logic in resize()
If the current size of the cache is larger than the new max size, we
call purge. It's not really necessary to check if the new size was
larger than the old size before checking the current size.

The advantage is that the test is now identical to setItem(), so it
can be generalized.